+++ sample/proto_v2/role_cfg.amd.js
@@ -13,7 +13,7 @@ define({
     ],
     data_ver : "1.0.0.0",
     hash_code : "sha256:eeaf27d61a221566a4129cce5f868e1d929d571f6ad2859de106242f0acc820f",
-    xres_ver : "2.17.1"
+    xres_ver : "2.18.0"
   },
   "role_cfg": [
     {