+++ sample/proto_v3/kind_option.js
@@ -41,6 +41,19 @@ exports.files = [
     message_type : {
       arr_in_arr : {
         field : {
+          int_arr : {
+            name : "int_arr",
+            number : 2,
+            options : {
+              field_list_max_size : 3,
+              field_list_min_size : 1,
+              field_list_strip_option : {
+                name : "LIST_STRIP_NOTHING",
+                number : 1
+              }
+            },
+            type_name : "INT32"
+          },
           name : {
             name : "name",
             number : 1,
@@ -91,6 +104,15 @@ exports.files = [
               field_separator : "|"
             },
             type_name : "MESSAGE"
+          },
+          test_plain_int_arr : {
+            name : "test_plain_int_arr",
+            number : 3,
+            options : {
+              field_list_max_size : 3,
+              field_list_min_size : 1
+            },
+            type_name : "INT32"
           },
           test_plain_msg : {
             name : "test_plain_msg",